requirements.txt
beautifulsoup4==4.14.3
playwright==1.57.0
psycopg2_binary==2.9.11
python-dotenv==1.2.1
Requests==2.32.5
stem==1.8.2
